Refactor common utils code
